last modification before publishing

+++ src/sender.h
@@ -1,7 +1,7 @@
 /*
  * Author : Antoine Gennart
  * Date : 2018-10
- * Description : This file is part of the project folder for the course 
+ * Description : This file is part of the project folder for the course
  *               LINGI1341 at UCLouvain.
  */
 
@@ -12,10 +12,10 @@
 /*
 	Create the next packet
 */
-pkt_status_code create_next_pkt(pkt_t* pkt, 
-								FILE *f, 
-								uint8_t pkt_nb, 
-								uint8_t window, 
+pkt_status_code create_next_pkt(pkt_t* pkt,
+								FILE *f,
+								uint32_t pkt_nb,
+								uint8_t window,
 								uint32_t nb_packet);
